refactor: ExceptionInfo 클래스 분리, 에러 메시지 상수 제거
